@@ -4,24 +4,27 @@
   "private": true,
   "basename": "/",
   "dependencies": {
-    "@babel/core": "7.4.3", 
+    "@babel/core": "7.4.3",
+    "@react-oauth/google": "^0.12.1",
     "@svgr/webpack": "4.1.0",
     "@typescript-eslint/eslint-plugin": "1.6.0",
     "@typescript-eslint/parser": "1.6.0",
-    "babel-eslint": "10.0.1",
     "axios": "^1.4.0",
+    "babel-eslint": "10.0.1",
     "babel-jest": "^24.9.0",
     "babel-loader": "8.0.5",
     "babel-plugin-named-asset-import": "^0.3.7",
     "babel-preset-react-app": "^9.1.2",
     "bootstrap": "^4.6.1",
     "camelcase": "^5.2.0",
+    "caniuse-lite": "^1.0.30001620",
     "case-sensitive-paths-webpack-plugin": "2.2.0",
     "css-loader": "2.1.1",
-    "dotenv": "6.2.0",
+    "dotenv": "^6.2.0",
     "dotenv-expand": "4.2.0",
     "emailjs-com": "^3.2.0",
     "eslint": "^5.16.0",
+    "eslint-config-google": "^0.14.0",
     "eslint-config-react-app": "^4.0.1",
     "eslint-loader": "2.1.2",
     "eslint-plugin-flowtype": "2.50.1",
@@ -30,6 +33,7 @@
     "eslint-plugin-react": "7.12.4",
     "eslint-plugin-react-hooks": "^1.7.0",
     "file-loader": "3.0.1",
+    "formik": "^2.4.6",
     "fs-extra": "7.0.1",
     "google-map-react": "^1.1.7",
     "html-webpack-plugin": "4.0.0-beta.5",
@@ -39,7 +43,9 @@
     "jest-environment-jsdom-fourteen": "0.1.0",
     "jest-resolve": "24.7.1",
     "jest-watch-typeahead": "0.3.0",
+    "jwt-decode": "^3.1.2",
     "mini-css-extract-plugin": "0.5.0",
+    "moment": "^2.30.1",
     "optimize-css-assets-webpack-plugin": "5.0.1",
     "pnp-webpack-plugin": "1.2.1",
     "postcss-flexbugs-fixes": "4.1.0",
@@ -47,28 +53,34 @@
     "postcss-normalize": "7.0.1",
     "postcss-preset-env": "6.6.0",
     "postcss-safe-parser": "4.0.1",
-    "react": "^17.0.2",
+    "prismjs": "^1.29.0",
+    "react": "^18.3.1",
     "react-accessible-accordion": "^3.3.5",
     "react-animate-on-scroll": "^2.1.5",
     "react-app-polyfill": "^1.0.6",
     "react-bootstrap": "^1.6.4",
     "react-countup": "^4.4.0",
+    "react-datepicker": "^6.9.0",
+    "react-datetime": "^3.2.0",
     "react-dev-utils": "^9.1.0",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.3.1",
+    "react-google-login": "^5.2.2",
     "react-helmet": "^5.2.1",
     "react-icons": "^3.11.0",
     "react-image-lightbox": "^5.1.4",
+    "react-js-pagination": "^3.0.3",
+    "react-modal": "^3.16.1",
     "react-modal-video": "^1.2.8",
     "react-parallax": "^3.3.0",
     "react-reveal": "^1.2.2",
-    "react-router-dom": "^5.3.0",
+    "react-router-dom": "^6.23.1",
     "react-scroll-up": "^1.3.7",
     "react-scrollspy": "^3.4.3",
     "react-slick": "^0.24.0",
     "react-tabs": "^3.2.3",
     "react-text-loop": "^2.3.0",
     "react-tilt": "^0.1.4",
-    "react-toastify": "^8.0.0",
+    "react-toastify": "^8.1.0",
     "react-tsparticles": "^1.39.3",
     "react-video-tag": "0.0.10",
     "react-visibility-sensor": "^5.1.1",
@@ -76,6 +88,7 @@
     "sass": "^1.50.1",
     "sass-loader": "^7.3.1",
     "semver": "6.0.0",
+    "sitemap": "^8.0.0",
     "slick-carousel": "^1.8.1",
     "style-loader": "0.23.1",
     "ts-pnp": "1.1.2",
@@ -83,7 +96,8 @@
     "webpack": "4.29.6",
     "webpack-dev-server": "3.2.1",
     "webpack-manifest-plugin": "2.0.4",
-    "workbox-webpack-plugin": "4.2.0"
+    "workbox-webpack-plugin": "4.2.0",
+    "yup": "^1.4.0"
   },
   "scripts": {
     "start": "node scripts/start.js",
@@ -156,6 +170,7 @@
     ]
   },
   "devDependencies": {
+    "@types/dotenv": "^8.2.0",
     "terser-webpack-plugin": "^1.4.5"
   }
 }
