@@ -1,15 +1,32 @@
 import React, { useState, useEffect } from "react";
 import { Link, useLocation } from "react-router-dom";
-import { FiX, FiMenu } from "react-icons/fi";
+import { FiX, FiMenu, FiUser, FiSun, FiMoon } from "react-icons/fi";
+import { useTheme } from "../../context/ThemeContext";
+// import { getCookie } from "../../auth/authUtils";
+import "./Header.css";
 
 const Header = (props) => {
     // State to hold the token
+    // const [token, setToken] = useState(getCookie('token'));
     const [token, setToken] = useState(localStorage.getItem('token'));
     const location = useLocation();
+    const { isDarkTheme, toggleTheme } = useTheme();
 
-    // Effect to update the token in state when it changes in local storage
+    // Effect to update the token in state when it changes in cookies
     useEffect(() => {
-        setToken(localStorage.getItem('token'));
+        const handleCookieChange = () => {
+            // setToken(getCookie('token'));
+            setToken(localStorage.getItem('token'));
+
+        };
+
+        // Listen to storage events to update the token state
+        window.addEventListener('storage', handleCookieChange);
+
+        // Cleanup listener on component unmount
+        return () => {
+            window.removeEventListener('storage', handleCookieChange);
+        };
     }, []);
 
     // Function to open the menu
@@ -24,7 +41,9 @@ const Header = (props) => {
 
     // Logout function
     const logout = () => {
-        localStorage.removeItem("token");
+        // document.cookie = 'token=; Max-Age=0; path=/; secure; samesite=strict';
+        localStorage.removeItem('token');
+        localStorage.removeItem('googleLogin');
         setToken(null);
         window.location.href = "/";
     };
@@ -41,19 +60,9 @@ const Header = (props) => {
 
     const { logo, color = 'default-color' } = props;
     let logoUrl;
-    if (logo === 'light') {
-        logoUrl = <img src="/assets/images/logo/logo-light.png" alt="Zeevoc" />;
-    } else if (logo === 'dark') {
-        logoUrl = <img src="/assets/images/logo/logo-dark.png" alt="Zeevoc" />;
-    } else if (logo === 'symbol-dark') {
-        logoUrl = <img src="/assets/images/logo/logo-symbol-dark.png" alt="Zeevoc" />;
-    } else if (logo === 'symbol-light') {
-        logoUrl = <img src="/assets/images/logo/logo-symbol-light.png" alt="Zeevoc" />;
-    } else {
-        logoUrl = <img src="/assets/images/logo/logo.png" alt="Zeevoc" />;
-    }
+    logoUrl = <img src="/assets/images/logo/logo.png" alt="Zeevoc" />;
 
-    return(
+    return (
         <header className={`header-area formobile-menu header--transparent ${color}`}>
             <div className="header-wrapper" id="header-wrapper">
                 <div className="header-left">
@@ -66,82 +75,38 @@ const Header = (props) => {
                 <div className="header-right">
                     <nav className="mainmenunav d-lg-block">
                         <ul className="mainmenu">
-                            {/* <li className="has-droupdown"><Link to="/">Home</Link>
-                                <ul className="submenu">
-                                    <li><Link to="/main-demo">Main Demo</Link></li>
-                                    <li><Link to="/dark-main-demo">Main Demo Dark</Link></li>
-                                    <li><Link to="/creative-agency">Creative Agency</Link></li>
-                                    <li><Link to="/creative-landing">Creative One Page</Link></li>
-                                    <li><Link to="/creative-portfolio">Creative Portfolio</Link></li>
-                                    <li><Link to="/personal-portfolio">Personal Portfolio</Link></li>
-                                    <li><Link to="/portfolio-landing">Portfolio One Page</Link></li>
-                                    <li><Link to="/dark-portfolio-landing">Portfolio One Page 02</Link></li>
-                                    <li><Link to="/digital-agency">Digital Agency</Link></li>
-                                    <li><Link to="/startup">Startup</Link></li>
-                                    <li><Link to="/paralax">Paralax</Link></li>
-                                    <li><Link to="/portfolio-home">Minimal Portfolio</Link></li>
-                                    <li><Link to="/business">Business</Link></li>
-                                    <li><Link to="/home-particles">Home Particles</Link></li>
-                                    <li><Link to="/studio-agency">Studio Agency</Link></li>
-                                    <li><Link to="/designer-portfolio">Designer Portfolio</Link></li>
-                                    <li><Link to="/interactive-agency">Interactive Agency</Link></li>
-                                </ul>
-                            </li> */}
                             <li><Link to="/" >Home</Link></li>
-                            {/* <li className="has-droupdown"><Link to="/service" >Service</Link>
-                                <ul className="submenu">
-                                    <li><Link to="/service">Service</Link></li>
-                                    <li><Link to="/service-details">Service Details</Link></li>
-                                </ul>
-                            </li>
-                            <li className="has-droupdown"><Link to="/service" >Blog</Link>
-                                <ul className="submenu">
-                                    <li><Link to="/blog">Blog List</Link></li>
-                                    <li><Link to="/blog-details">Blog Details</Link></li>
-                                </ul>
-                            </li>
-                            <li className="has-droupdown"><Link to="#pages" >Portfolio</Link>
-                                <ul className="submenu">
-                                    <li><Link to="/portfolio">Portfolio</Link></li>
-                                    <li><Link to="/portfolio-details">Portfolio Details</Link></li>
-                                    <li><Link to="/404">404</Link></li>
-                                </ul>
-                            </li> */}
-                            {/* <li className="has-droupdown"><Link to="#" >Blocks</Link>
-                                <ul className="submenu">
-                                    <li><Link to="/portfolio">Portfolio</Link></li>
-                                    <li><Link to="/team">Team</Link></li>
-                                    <li><Link to="/service">Service</Link></li>
-                                    <li><Link to="/video-popup">Video Popup</Link></li>
-                                    <li><Link to="/progressbar">Progressbar</Link></li>
-                                    <li><Link to="/gallery">Gallery</Link></li>
-                                    <li><Link to="/counters">Counters</Link></li>
-                                    <li><Link to="/blog">Blog List</Link></li>
-                                    <li><Link to="/clint-logo">Clint Logo</Link></li>
-                                    <li><Link to="/contact-form">Contact Form</Link></li>
-                                    <li><Link to="/google-map">Google Map</Link></li>
-                                    <li><Link to="/columns">Columns</Link></li>
-                                    <li><Link to="/pricing-table">Pricing Table</Link></li>
-                                </ul>
-                            </li> */}
-                            <li><Link to="/service" >Service</Link></li>
-                            <li><Link to="/blog" >Blog</Link></li>
-                            {/* <li><Link to="/portfolio" >Portfolio</Link></li> */}
-                            {/* <li><Link to="/products" >Products</Link></li> */}
+                            <li><Link to="/services" >Services</Link></li>
+                            <li><Link to="/blogs" >Blogs</Link></li>
+                            <li><Link to="/extensions" >Extensions</Link></li>
                             <li><Link to="/contact" >Contact</Link></li>
                             <li><Link to="/about" >About</Link></li>
-                            {/* {(location.pathname !== '/login' && !token) && <li><Link to="/login">Login</Link></li>} */}
-                            {/* {(location.pathname !== '/signup' && !token) && <li><Link to="/signup">Sign Up</Link></li>} */}
-                            {/* {token && <li onClick={logout} > <Link >Logout</Link></li>} */}
-                            {/* <li><Link to="/signup" >Sign Up</Link></li> */}
-                            {/* <li><Link to="/signup" >Sign Up</Link></li> */}
+                            <li>
+                                <Link to="#" onClick={toggleTheme} className="theme-toggle-icon">
+                                    {isDarkTheme ? "Theme " : "Theme "}
+                                </Link>
+                            </li>
+                            {location.pathname !== '/login' && (
+                                <li className="has-droupdown">
+                                    <Link to="#">
+                                        Account
+                                        {/* <FiUser /> */}
+                                    </Link>
+                                    <ul className="submenu">
+                                        {/* Theme Toggle Icon */}
+
+                                        {token && <li><Link to="/orders">Orders</Link></li>}
+                                        {token && <li><Link to="/profile">Profile</Link></li>}
+                                        {(location.pathname !== '/login' && !token) && <li><Link to="/login">Login</Link></li>}
+                                        {token && <li onClick={logout}><Link to="#">Logout</Link></li>}
+                                    </ul>
+                                </li>
+                            )}
+
+
                         </ul>
                     </nav>
-                    {/* <div className="header-btn">
-                        <a className="rn-btn" href="/signup">
-                            <span>Sign Up</span>
-                        </a>
-                    </div> */}
+
                     {/* Start Humberger Menu  */}
                     <div className="humberger-menu d-block d-lg-none pl--20">
                         <span onClick={menuTrigger} className="menutrigger text-white"><FiMenu /></span>
