@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import Header from "../component/header/HeaderFour";
+import Header from "../component/header/Header";
 import ScrollToTop from 'react-scroll-up';
 import { FiChevronUp } from "react-icons/fi";
 import Footer from "../component/footer/FooterTwo";
