@@ -1,68 +1,53 @@
-import React, { Component } from "react";
-import {Link} from "react-router-dom";
+// import React, { useState, useEffect } from "react";
+// import { Link } from "react-router-dom";
+// import { getAllProducts } from "../../api";
 
-const PortfolioListContent = [
-    {
-        image: 'image-1',
-        category: 'Magento 2 Extension',
-        title: 'Getting tickets to the big show'
-    },
-    {
-        image: 'image-2',
-        category: 'Magento 2 Extension',
-        title: 'Getting tickets to the big show'
-    },
-    {
-        image: 'image-3',
-        category: 'Magento 2 Extension',
-        title: 'Getting tickets to the big show'
-    },
-    {
-        image: 'image-4',
-        category: 'Magento 2 Extension',
-        title: 'Getting tickets to the big show'
-    },
-    {
-        image: 'image-3',
-        category: 'Magento 2 Extension',
-        title: 'Getting tickets to the big show'
-    },
-    {
-        image: 'image-4',
-        category: 'Magento 2 Extension',
-        title: 'Getting tickets to the big show'
-    }
-]
+// const PortfolioList = ({ column, styevariation, item }) => {
+//   const [products, setProducts] = useState([]);
 
-class PortfolioList extends Component{
-    render(){
-        const {column , styevariation } = this.props;
-        const list = PortfolioListContent.slice(0 , this.props.item);
-        return(
-            <React.Fragment> 
-                {list.map((value , index) => (
-                    <div className={`${column}`} key={index}>
-                        <div className={`portfolio ${styevariation}`}>
-                            <div className="thumbnail-inner">
-                                <div className={`thumbnail ${value.image}`}></div>
-                                <div className={`bg-blr-image ${value.image}`}></div>
-                            </div>
-                            <div className="content">
-                                <div className="inner">
-                                    <p>{value.category}</p>
-                                    <h4><a href="/portfolio-details">{value.title}</a></h4>
-                                    <div className="portfolio-button">
-                                        <a className="rn-btn" href="/portfolio-details">Buy Now</a>
-                                    </div>
-                                </div>
-                            </div>
-                            <Link className="link-overlay" to="/portfolio-details"></Link>
-                        </div>
-                    </div>
-                ))}
-               
-            </React.Fragment>
-        )
-    }
-}
-export default PortfolioList;
+//   useEffect(() => {
+//     const fetchProducts = async () => {
+//       try {
+//         const response = await getAllProducts();
+//         setProducts(response.data);
+//       } catch (error) {
+//         console.error("Error fetching products:", error);
+//       }
+//     };
+
+//     fetchProducts();
+//   }, []);
+
+//   const productList = products.slice(0, item);
+
+//   return (
+//     <React.Fragment>
+//       {productList.map((product, index) => (
+//         <div className={`${column}`} key={index}>
+//           <div className={`portfolio ${styevariation}`}>
+//             <div className="thumbnail-inner">
+//               <div className={`thumbnail ${product.image}`}></div>
+//               <div className={`bg-blr-image ${product.image}`}></div>
+//             </div>
+//             <div className="content">
+//               <div className="inner">
+//                 <p>{product.category}</p>
+//                 <h4>
+//                   <Link to={`/product/${product._id}`}>{product.title}</Link>
+//                 </h4>
+//                 <div className="portfolio-button">
+//                   <Link className="rn-btn" to={`/product/${product._id}`}>
+//                     Buy Now
+//                   </Link>
+//                 </div>
+//               </div>
+//             </div>
+//             <Link className="link-overlay" to={`/product/${product._id}`}></Link>
+//           </div>
+//         </div>
+//       ))}
+//     </React.Fragment>
+//   );
+// };
+
+// export default PortfolioList;
