@@ -1,12 +1,7 @@
 import React, { useState, useEffect } from "react";
-import {
-    FiCode,
-    FiBox,
-    FiShoppingCart,
-    FiBookOpen,
-    FiSettings,
-    FiGlobe
-} from "react-icons/fi";
+import { FiChevronUp, FiShoppingBag, FiShoppingCart, FiBookOpen, FiSettings, FiMonitor, FiCode } from "react-icons/fi";
+import LoadingSpinner from "../../component/spinner/LoadingSpinner";
+
 import { getAllServices } from "../../api";
 
 const ServiceTwo = () => {
@@ -18,7 +13,6 @@ const ServiceTwo = () => {
             try {
                 const response = await getAllServices();
                 setServices(response.data);
-                console.log(response);
                 setLoading(false);
             } catch (error) {
                 console.error("Error fetching services:", error);
@@ -30,7 +24,7 @@ const ServiceTwo = () => {
     }, []);
 
     let title = 'Services',
-        description = 'Don\'t fall for the lies! Our transparent approach ensures honesty and integrity in every aspect of your business.';
+        description = 'Don\'t fall for the lies Our transparent approach ensures honesty and integrity in every aspect of your business.';
 
     return (
         <React.Fragment>
@@ -40,26 +34,25 @@ const ServiceTwo = () => {
                         <h2 className="title">{title}</h2>
                         <p>{description}</p>
                         <div className="service-btn">
-                            <a className="btn-transparent rn-btn-dark" href="/service"><span className="text">Request Custom Service</span></a>
+                            <a className="btn-transparent rn-btn-dark" href="/services"><span className="text">View All Services</span></a>
                         </div>
                     </div>
                 </div>
                 <div className="col-lg-8 col-12 mt_md--50">
                     <div className="row service-one-wrapper">
-                        {loading ? (
-                            <p>Loading...</p>
+                        {loading? (
+                           <LoadingSpinner/>
                         ) : (
-                            services.map((service, index) => (
+                            services.slice(0, 4).map((service, index) => (
                                 <div className="col-lg-6 col-md-6 col-sm-6 col-12" key={index}>
-                                    <a href="/service-details">
+                                    {/* <a href={`/service/${service.urlKey}`}> */}
+                                    <a>
                                         <div className="service service__style--2">
                                             <div className="icon">
-                                            {index === 0 && <FiCode />}
-                                            {index === 1 && <FiBox />}
-                                            {index === 2 && <FiShoppingCart />}
-                                            {index === 3 && <FiBookOpen />}
-                                            {index === 4 && <FiSettings />}
-                                            {index === 5 && <FiGlobe />}
+                                                {index === 0 && <FiShoppingBag />} {/* Magento Development */}
+                                                {index === 1 && <FiCode />} {/* Extension Development */}
+                                                {index === 2 && <FiShoppingCart />} {/* E-commerce Solutions */}
+                                                {index === 3 && <FiBookOpen />} {/* Academic Project Assistance */}
                                             </div>
                                             <div className="content">
                                                 <h3 className="title">{service.title}</h3>
