@@ -1,5 +1,5 @@
 import express from "express";
-import auth from "../middleware/auth.js";
+import { adminAuth } from "../middleware/auth.js";
 
 import {
   createBlogPost,
@@ -7,14 +7,20 @@ import {
   getBlogPostById,
   updateBlogPost,
   deleteBlogPost,
+  getBlogByUrlKey,
+  createMultipleBlogPosts,
+  deleteAllBlogPosts
 } from "../controllers/blogPost.js";
 
 const router = express.Router();
 
-router.post("/create", createBlogPost);
+router.post("/create", adminAuth, createBlogPost);
 router.get("/all", getAllBlogPosts);
 router.get("/getbyid/:id", getBlogPostById);
-router.patch("/updatebyid/:id", updateBlogPost);
-router.delete("/deletebyid/:id", deleteBlogPost);
+router.get("/:urlKey", getBlogByUrlKey);
+router.patch("/updatebyid/:id", adminAuth, updateBlogPost);
+router.delete("/deletebyid/:id", adminAuth, deleteBlogPost);
+router.delete("/deleteall", adminAuth, deleteAllBlogPosts);
+router.post('/createmultiple', createMultipleBlogPosts);
 
 export default router;
